Handle internet dropouts

+++ source/main.py
@@ -20,8 +20,8 @@ def load_all_posts():
 
 
 def update_posts():
-    # this won't record 100 new posts per day, it will record all the posts that are above the minimum upvotes
-    post_recorder.record_new_posts('day', 100)
+    # this won't record 200 new posts per day, it will record all the posts that are above the minimum upvotes
+    post_recorder.record_new_posts('day', 200)
 
     # refresh image_searcher to load in the updated index.csv file
     global image_searcher
@@ -32,11 +32,13 @@ def update_posts():
 
 
 def reply(repost: NewPost, original):
+    if repost.link in replied: return
+
     with open(bot_activity, 'a') as file:
         file.write('{} is a repost of {}\n'.format(repost.link, original.link))
     message = config.message.format(original.title, original.link)
-    repost.submission.reply(message)
-    print("Replied to: {}".format(repost.link))
+    # repost.submission.reply(message)
+    replied.append(repost.link)
 
 
 def check_if_repost(new_post):
@@ -69,38 +71,46 @@ def check_if_repost(new_post):
 
 
 def main():
-    if os.path.exists(new_folder):
-        shutil.rmtree(new_folder)
     process_start_time = datetime.today()
 
-    for submission in subreddit.stream.submissions():
-        post_age = datetime.today().timestamp() - submission.created_utc
-        if post_age > config.max_post_age: continue  # so the bot can be restarted without analysing a backlog of posts
-        print(post_age)
-
-        post = NewPost(submission, new_folder)
-
-        # handle internet dropouts
+    # this is wrapped in a try except to prevent internet dropouts crashing the bot
+    while True:
         try:
-            check_if_repost(post)
+            for submission in subreddit.stream.submissions():
+                post_age = datetime.today().timestamp() - submission.created_utc
+                if post_age > config.max_post_age: continue  # so the bot can be restarted without analysing old posts
+                print(post_age)
+
+                post = NewPost(submission, new_folder)
+                check_if_repost(post)
+                shutil.rmtree(new_folder)
+
+                # update posts when there is low traffic
+                if process_start_time.day < datetime.today().day and datetime.today().hour == config.update_hour:
+                    print('\n refreshing posts at {0:02d}'.format(datetime.today().hour)
+                          + ':{0:02d}'.format(datetime.today().minute))
+
+                    global posts
+                    posts = update_posts()
+                    process_start_time = datetime.today()
         except prawcore.exceptions.ServerError:
-            time.sleep(180)
-
-        shutil.rmtree(new_folder)
-
-        # update posts when there is low traffic
-        if process_start_time.day < datetime.today().day and datetime.today().hour == config.update_hour:
-            print('\n refreshing posts at {0:02d}'.format(datetime.today().hour)
-                  + ':{0:02d}'.format(datetime.today().minute))
-
-            global posts
-            posts = update_posts()
-            process_start_time = datetime.today()
+            print("Error occured while connecting to reddit, restarting stream")
+            time.sleep(300)
+        except Exception as e:
+            print("\n\n\nError occured: {}".format(e))
+            print("If this isn't an internet connection error please report it "
+                  "at https://github.com/Wattsy2020/repost-detector/issues\n\n\n")
+            time.sleep(300)
 
 
 new_folder = os.path.join(post_recorder.base_folder, 'new')
 temp_folder = os.path.join(post_recorder.base_folder, 'temp')
+
+if os.path.exists(new_folder): shutil.rmtree(new_folder)
+if os.path.exists(temp_folder): shutil.rmtree(temp_folder)
+
 bot_activity = os.path.join(post_recorder.base_folder, 'reposts.txt')
+replied = []  # track replied too posts, if an error occurs and the bot has to restart it won't reply to a post twice
 
 posts = load_all_posts()
 image_searcher = image_search.ImageSearcher(post_recorder.index_file)
